#13: Refactoring
