--- jfxui/src/main/java/org/itsallcode/whiterabbit/jfxui/uistate/WidgetState.java
@@ -1,8 +1,0 @@
-package org.itsallcode.whiterabbit.jfxui.uistate;
-
-interface WidgetState<T>
-{
-    void store(T widget);
-
-    void restore(T widget);
-}
