+++ jfxui/src/main/java/org/itsallcode/whiterabbit/jfxui/uistate/widgets/WidgetState.java
@@ -1,0 +1,8 @@
+package org.itsallcode.whiterabbit.jfxui.uistate.widgets;
+
+public interface WidgetState<T>
+{
+    void store(T widget);
+
+    void restore(T widget);
+}
